serialise_workflow_exception in helpers
